Move checkConnectionCode utility function to a common place
