+++ projects/js-packages/publicize-components/src/utils/connections.ts
@@ -1,4 +1,4 @@
-import { Connection } from '../../social-store/types';
+import { Connection } from '../social-store/types';
 
 export const checkConnectionCode = ( connection: Connection, code: string ) => {
 	return false === connection.is_healthy && code === ( connection.error_code ?? 'broken' );
